project was restructured clean arch for middlewares check that staff is owner in Add function in usecases added
