+++ go.mod
@@ -1,0 +1,20 @@
+module 2020_1_drop_table
+
+go 1.13
+
+require (
+	github.com/garyburd/redigo v1.6.0 // indirect
+	github.com/go-playground/locales v0.13.0
+	github.com/go-playground/universal-translator v0.17.0
+	github.com/gorilla/mux v1.7.4
+	github.com/gorilla/sessions v1.2.0
+	github.com/jmoiron/sqlx v1.2.0
+	github.com/leodido/go-urn v1.2.0 // indirect
+	github.com/lib/pq v1.3.0
+	github.com/nu7hatch/gouuid v0.0.0-20131221200532-179d4d0c4d8d
+	github.com/rs/zerolog v1.18.0
+	github.com/skip2/go-qrcode v0.0.0-20191027152451-9434209cb086
+	github.com/stretchr/testify v1.5.1
+	gopkg.in/boj/redistore.v1 v1.0.0-20160128113310-fc113767cd6b
+	gopkg.in/go-playground/validator.v9 v9.31.0
+)
