+++ internal/app/models/errors.go
