+++ internal/pkg/qr/qr.go
