--- projectConfig/cofigFile.go
@@ -1,4 +1,0 @@
-package projectConfig
-
-const MediaFolder = "media"
-const ServerUrl = "http://localhost:8080"
